Warframe Damage Quantization / Shattering Impact Armor Stripping
I'm testing Warframes Damage Quantization as i noticed a not in the wiki which said it wouln't work on the Galantine. After a few tests i found out it does work unless you bring slash damage mods, now this ingame logic is at war with that the wiki says and i would like to figure out whats going wrong.